[xmodmap] Rebind key of Rapoo_V500

+++ install.sh
@@ -2,57 +2,58 @@
 
 DIR=$(cd $(dirname $0) && pwd)
 
-declare -A CONFIG_FUNC=(
-  [1]="config_bashrc"
-  [2]="config_bash_profile"
-  [3]="config .vimrc .vim"
-  [4]="config .tmux.conf"
-  [5]="config .gitconfig .git-extensions"
-  [6]="config_xmodmap"
-  [7]="config .gnupg/gpg.conf .gnupg/dirmngr.conf .gnupg/sks-keyserver.netCA.pem"
-  [8]="config .w3m/keymap"
-  [9]="config_suckless dwm"
-  [10]="config_suckless st"
-  [11]="config_suckless slstatus"
-  [12]="config_suckless dmenu"
-  [13]="config_suckless slock"
-  [14]="config .config/dunst"
-  [15]="config .xinitrc"
-  [16]="config .infokey"
-  [17]="config_pam_environment"
-)
-
-declare -A CONFIG_CB=(
-  [1]="echo Please run the command \'source \~/.bashrc\'"
-  [2]=""
-  [3]="callback_vim"
-  [4]=""
-  [5]=""
-  [6]="xmodmap ~/.Xmodmap"
-  [7]=""
-  [8]=""
-  [9]=""
-  [10]=""
-  [11]=""
-  [12]=""
-  [13]=""
-  [14]=""
-  [15]=""
-  [16]=""
+declare -A PROGRAMS=(
+  [xinitrc]="
+    config .xinitrc"
+  [xmodmap]="
+    config_xmodmap
+      && xmodmap ~/.xmodmaprc"
+  [bashrc]="
+    config .bashrc
+      && echo Please run the command 'source ~/.bashrc'."
+  [bash_profile]="
+    config_bash_profile"
+  [info]="
+    config .infokey"
+  [vim]="
+    config .vimrc .vim
+      && post_vim"
+  [tmux]="
+    config .tmux.conf"
+  [git]="
+    config .gitconfig .git-extensions"
+  [gpg]="
+    config .gnupg/gpg.conf .gnupg/dirmngr.conf .gnupg/sks-keyserver.netCA.pem"
+  [dunst]="
+    config .config/dunst
+      && post_dunst"
+  #[w3m]="config .w3m/keymap"
+  #[dwm]="config_suckless dwm"
+  #[st]="config_suckless st"
+  #[slstatus]="config_suckless slstatus"
+  #[dmenu]="config_suckless dmenu"
+  #[slock]="config_suckless slock"
+  #[pam]="config_pam_environment"
 )
 
 _backup() {
   # :param $1: file to be backed up
 
+  # When file doesn't exist
   [[ -e "${1}" ]] || { echo "Create new file '$1'"; return; }
-  [[ -h "${1}" ]] && rm -vf "$1" && return
+  # When file is a symbolic link
+  [[ -h "${1}" ]] && { rm -vf "$1"; return; }
+
+  local backup
+  read -e -p "Do you want to backup original configuration files? (y/n): " backup
+  [[ ${backup} =~ [Nn] ]] && return
 
   local dotbak
   local suffix=0
   
   for dotbak in $(ls ${1}.bak.[0-9] ${1}.bak.[1-9][0-9]* 2>/dev/null); do
     (( suffix < ${dotbak##*.} )) && suffix=${dotbak##*.}
-    # compare only when they are regular files
+    # Compare them only when they are regular files
     [[ -f "${1}" && -f "${dotbak}" ]] && {
       cmp --silent "${1}" "${dotbak}" && rm -f "${dotbak}"
     }
@@ -88,22 +89,24 @@ _write_cmdlines() {
   done > ${1}
 }
 
-config_bashrc() {
-  _read_cmdlines ~/.bashrc
-
-  cmdlines+=(
-    '# werice start {{{'
-    '[[ -f ~/.me/merc ]] && . ~/.me/merc'
-    '# }}} werice end'
-  )
-
-  [[ ${backup} =~ [Nn] ]] || _backup ~/.me
-  ln -vsf ${DIR}/.me ~/.me
-
-  [[ "${cmdlines[*]}" == "${cmdlines_old[*]}" ]] && return
-  [[ ${backup} =~ [Nn] ]] || _backup ~/.bashrc
-  _write_cmdlines ~/.bashrc
-}
+#config_bashrc() {
+#  _read_cmdlines ~/.bashrc
+#
+#  cmdlines+=(
+#    '# werice start {{{'
+#    '[[ -f ~/.me/merc ]] && . ~/.me/merc'
+#    '# }}} werice end'
+#  )
+#
+#  #[[ ${backup} =~ [Nn] ]] || _backup ~/.me
+#  _backup ~/.me
+#  ln -vsf ${DIR}/.me ~/.me
+#
+#  [[ "${cmdlines[*]}" == "${cmdlines_old[*]}" ]] && return
+#  #[[ ${backup} =~ [Nn] ]] || _backup ~/.bashrc
+#  _backup ~/.bashrc
+#  _write_cmdlines ~/.bashrc
+#}
 
 config_bash_profile() {
   _read_cmdlines ~/.bash_profile
@@ -116,7 +119,8 @@ config_bash_profile() {
     '# }}} werice end'
   )
 
-  [[ ${backup} =~ [Nn] ]] || _backup ~/.bash_profile
+  #[[ ${backup} =~ [Nn] ]] || _backup ~/.bash_profile
+  _backup ~/.bash_profile
   _write_cmdlines ~/.bash_profile
 }
 
@@ -131,7 +135,8 @@ config_pam_environment() {
     '# }}} werice end'
   )
 
-  [[ ${backup} =~ [Nn] ]] || _backup ~/.pam_environment
+  #[[ ${backup} =~ [Nn] ]] || _backup ~/.pam_environment
+  _backup ~/.pam_environment
   _write_cmdlines ~/.pam_environment
 }
 
@@ -142,94 +147,72 @@ config_xmodmap() {
 2. HP EliteBook 8470p
 3. ACER TravelMate TX50
 EOF
-  read -p "Which keyboard do you want to set? (1/2): " choice
+  read -e -p "Which keyboard do you want to set? (1/2/3): " choice
   case $choice in
-    1) ln -sf ${DIR}/xmodmap/Rapoo_V500.Xmodmap ~/.Xmodmap ;;
-    2) ln -sf ${DIR}/xmodmap/HP_EliteBook_8470p.Xmodmap ~/.Xmodmap ;;
-    3) ln -sf ${DIR}/xmodmap/ACER_TravelMate_TX50.Xmodmap ~/.Xmodmap ;;
+    1) ln -sf ${DIR}/xmodmap/Rapoo_V500.xmodmaprc ~/.xmodmaprc ;;
+    2) ln -sf ${DIR}/xmodmap/HP_EliteBook_8470p.xmodmaprc ~/.xmodmaprc ;;
+    3) ln -sf ${DIR}/xmodmap/ACER_TravelMate_TX50.xmodmaprc ~/.xmodmaprc ;;
     *) echo "Unknown option." ;;
   esac
 }
 
-config_suckless() {
-  # :param $1: suckless module to be installed
-
-  [[ -h ~/.suckless ]] || ln -sf ${DIR}/suckless ~/.suckless
-  [[ $(hostname) =~ ^peace|cheese$ ]] || {
-    echo "Unkonwn host (only support peace or cheese)."
-    exit 1
-  }
-
-  local MOD_DIR=${DIR}/suckless/$1
-  git submodule update --init $MOD_DIR
-
-  set -i
-  cd ${MOD_DIR}
-  git stash
-  patch -b -o config.h config.def.h \
-    $MOD_DIR-patches/$1-config$(hostname)-*-$(git rev-parse --short HEAD).diff
-  git apply \
-    $MOD_DIR-patches/$1-custom-*-$(git rev-parse --short HEAD).diff
-  make && sudo make install && make clean
-  cd ${DIR}
-}
+#config_suckless() {
+#  # :param $1: suckless module to be installed
+#
+#  [[ -h ~/.suckless ]] || ln -sf ${DIR}/suckless ~/.suckless
+#  [[ $(hostname) =~ ^peace|cheese$ ]] || {
+#    echo "Unkonwn host (only support peace or cheese)."
+#    exit 1
+#  }
+#
+#  local MOD_DIR=${DIR}/suckless/$1
+#  git submodule update --init $MOD_DIR
+#
+#  set -i
+#  cd ${MOD_DIR}
+#  git stash
+#  patch -b -o config.h config.def.h \
+#    $MOD_DIR-patches/$1-config$(hostname)-*-$(git rev-parse --short HEAD).diff
+#  git apply \
+#    $MOD_DIR-patches/$1-custom-*-$(git rev-parse --short HEAD).diff
+#  make && sudo make install && make clean
+#  cd ${DIR}
+#}
 
 config() {
   # :param $@: configuration file to be installed
 
   for file in ${@}; do
     mkdir -pv $(dirname ${file})
-    [[ ${backup} =~ [Nn] ]] || _backup ~/${file}
+    #[[ ${backup} =~ [Nn] ]] || _backup ~/${file}
+    _backup ~/"$file"
     ln -vsf ${DIR}/${file} ~/${file}
   done
 }
 
-callback_vim() {
+post_vim() {
   # vim bundle
   vim -e -c "call Bundle('install') | visual | qa"
+}
+
+post_dunst() {
+  systemctl --user import-environment DISPLAY
+  systemctl --user start dunst
+  systemctl --user enable dunst
 }
 
 
 # Main
 cat <<EOF
                        === rice installation ===
-This script is intend to install configuration files for the following programs.
-  0. all
-  1. bashrc
-  2. bash_profile
-  3. vim
-  4. tmux
-  5. git
-  6. xmodmap
-  7. gpg
-  8. w3m
-  9. dwm
- 10. st
- 11. slstatus
- 12. dmenu
- 13. slock
- 14. dunst
- 15. xinitrc
- 16. info
- 17. pam_environment
+This script is intend to install configuration files for the following programs:
+${!PROGRAMS[*]}
 EOF
-read -p "Which configuration file do you want to install? (0/1/.../${#CONFIG_FUNC[@]}): " choice
-read -p "Do you want to backup your original configurations? (y/n): " backup
+read -e -p "Which configuration file of program do you want to install? : " choice
 echo ""
 
-case ${choice} in
-  0)
-    for (( i = 1; i <= ${#CONFIG_FUNC[@]}; i++ )); do
-      eval "${CONFIG_FUNC[${i}]}"
-      eval "${CONFIG_CB[${i}]}"
-    done ;;
-  *)
-    if [[ -n ${CONFIG_FUNC[${choice}]} ]]; then
-      eval "${CONFIG_FUNC[${choice}]}"
-      eval "${CONFIG_CB[${choice}]}"
-    else
-      echo "Unknown choice of configuration files :("
-      exit 1
-    fi ;;
-esac
-
+if [[ $choice =~ $(IFS='|'; echo "${!PROGRAMS[*]}") ]]; then
+  eval "${PROGRAMS[$choice]}"
+else
+  echo "Unknown choice of configuration files :("
+fi
